refactor: CLIテストの *TestdataPath ヘルパーを testhelper_test.go に共通化する
config_check_test.go / feedgen_check_test.go / episodegen_check_test.go /
assets_test.go に同一実装の TestdataPath ヘルパーが重複していたため、
testhelper_test.go を新設して cliTestSrcDir・init・configTestdataPath を
一箇所に集約した。各ファイルの重複定義は削除し、共通の configTestdataPath を呼ぶよう統一した。
