+++ internal/cli/config_check_test.go
@@ -2,25 +2,11 @@ package cli_test
 
 import (
 	"bytes"
-	"path/filepath"
-	"runtime"
 	"strings"
 	"testing"
 
 	"github.com/canpok1/vox-radio/internal/cli"
 )
-
-// cliTestSrcDir is the absolute path of this test file's directory, resolved at init time.
-var cliTestSrcDir string
-
-func init() {
-	_, file, _, _ := runtime.Caller(0)
-	cliTestSrcDir = filepath.Dir(file)
-}
-
-func configTestdataPath(rel string) string {
-	return filepath.Join(cliTestSrcDir, "..", "config", "testdata", rel)
-}
 
 func TestConfigCheck_ValidYAML_Success(t *testing.T) {
 	cmd := cli.NewRootCmd()
